chore: extract shared code between regionizers
